Revert useless POI classes changes
