+++ src/adapters/poi/poi_store.js
@@ -2,7 +2,6 @@ import Poi from './poi';
 import Store from '../../adapters/store';
 import Error from '../error';
 import Telemetry from '../../libs/telemetry';
-import IdunnPoi from './idunn_poi';
 
 const store = new Store();
 export default class PoiStore extends Poi {
@@ -30,17 +29,5 @@ export default class PoiStore extends Poi {
     return storedData.map(poi => {
       return Object.assign(new PoiStore(), poi);
     });
-  }
-
-  static deserialize(raw) {
-    const {
-      id, name, alternativeName, type, latLon, className, subClassName, bbox, blocks,
-      topImageUrl, kind,
-    } = raw;
-    if (kind === 'idunn') {
-      return new IdunnPoi(raw);
-    }
-    return new Poi(id, name, alternativeName, type, latLon, className, subClassName, bbox, blocks,
-      topImageUrl, raw);
   }
 }
